requirements.txt
appdirs==1.4.0
Babel==2.4.0
bitarray==0.8.1
cuckoopy==0.1.1
daemon==1.1
debtcollector==1.13.0
elasticsearch==5.2.0
funcsigs==1.0.2
futures==3.0.5
get==0.0.0
iso8601==0.1.11
jsonify==0.5
keystoneauth1==2.19.0
mmh3==2.3.1
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.5
oslo.config==3.24.0
oslo.i18n==3.15.0
oslo.serialization==2.18.0
oslo.utils==3.25.0
packaging==16.8
pbr==2.0.0
positional==1.1.1
post==0.0.0
public==0.0.0
py==1.4.33
pyparsing==2.1.10
pytest==3.0.7
python-keystoneclient==3.10.0
python-swiftclient==3.3.0
pytz==2017.2
-e git+https://github.com/greglange/qork.git@ec536dafbd94b5e39954f81f1929c23184391900#egg=qork
query-string==0.0.0
request==0.0.0
requests==2.13.0
rfc3986==0.4.1
sarge==0.1.4
setupfiles==0.0.0
six==1.10.0
stevedore==1.21.0
swift==1.0.2
update==0.4.4
urllib3==1.20
wrapt==1.10.10
